pyproject.toml
[project]
name = "brighter-bookings-ms"
version = "0.1.0"
description = "Bookings microservice for the BrighterProject platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.112.4",
    "httpx>=0.28",
    "loguru>=0.7.3",
    "ms-core",
    "pydantic>=2.11.7",
    "redis>=7.2.0",
    "slowapi>=0.1.9",
    "tortoise-orm[asyncpg]>=1.1.7",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.9",
    "taskipy>=1.14",
]

[tool.taskipy.tasks]
check = "bash scripts/check.sh"

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]  # Depends() in FastAPI function defaults is intentional

[tool.pytest.ini_options]
addopts = "-v -m 'not integration'"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests that require a live Redis instance (deselected by default; run with -m integration)",
]

[tool.pyright]
reportInvalidTypeForm = "none"

[tool.uv.sources]
ms-core = { url = "https://github.com/HexChap/MSCore/archive/refs/heads/master.zip" }

[tool.tortoise]
tortoise_orm = "main.TORTOISE_ORM"
